Add homepage content to production seeds
Refactoring the logger logic to a separate file and including it in
`dev_seeds` and `seeds` as the homepage seeds[1] use the logger method
`section`

https://github.com/consul/consul/blob/master/db/dev_seeds/widgets.rb#L1
