+++ db/dev_seeds.rb
@@ -1,20 +1,7 @@
 require 'database_cleaner'
 DatabaseCleaner.clean_with :truncation
-@logger = Logger.new(STDOUT)
-@logger.formatter = proc do |_severity, _datetime, _progname, msg|
-                      msg unless @avoid_log
-                    end
 
-def section(section_title)
-  @logger.info section_title
-  yield
-  log(' ✅')
-end
-
-def log(msg)
-  @logger.info "#{msg}\n"
-end
-
+require_relative 'seed_logger'
 require_relative 'dev_seeds/settings'
 require_relative 'dev_seeds/geozones'
 require_relative 'dev_seeds/users'
