+++ db/seed_logger.rb
@@ -1,0 +1,14 @@
+@logger = Logger.new(STDOUT)
+@logger.formatter = proc do |_severity, _datetime, _progname, msg|
+                      msg unless @avoid_log
+                    end
+
+def section(section_title)
+  @logger.info section_title
+  yield
+  log(' ✅')
+end
+
+def log(msg)
+  @logger.info "#{msg}\n"
+end
